Rename __main__.py -> codemodder.py. Remove callable module
