+++ tests/samples/django-project/mysite/manage.py
@@ -18,5 +18,5 @@ def main():
     execute_from_command_line(sys.argv)
 
 
-if __name__ == "__main__":
+if __name__ == "codemodder":
     main()
